a bare minimum kb class just to hold db connections
