+++ ccm_benchmate/knowledge_base/utils.py
@@ -1,27 +1,2 @@
-#TODO
-#import
-
-# spectre2
-def embed_chunk(text, model):
-    pass
-
-def chunk_text(text):
-    pass
-
-#this can be an image of a figure or a table copali
-def embed_image(image, model):
-    pass
-
-def keyword_search(database, table, **kargs):
-    pass
-
-#colbert re-ranking
-def re_rank(results, ranking_model):
-    pass
-
-def retrieve(database, table, **kargs):
-    pass
-
-
